Translate a numeric vector to a cardinal character vector

Convert a numeric vector to a cardinal numeral (e.g. one tenth, one, two).

numeric_friendly_safe() checks that all arguments are of the correct type and raises an informative error otherwise. numeric_friendly() does not perform input validation to maximize its speed.

Usage

numeric_friendly(
  numbers,
  zero = "zero",
  na = "missing",
  nan = "not a number",
  inf = "infinity",
  negative = "negative ",
  decimal = " and ",
  and = FALSE,
  hyphenate = TRUE,
  and_fractional = and,
  hyphenate_fractional = hyphenate,
  english_fractions = NULL
)

numeric_friendly_safe(
  numbers,
  zero = "zero",
  na = "missing",
  nan = "not a number",
  inf = "infinity",
  negative = "negative ",
  decimal = " and ",
  and = FALSE,
  hyphenate = TRUE,
  and_fractional = and,
  hyphenate_fractional = hyphenate,
  english_fractions = NULL
)

Arguments

numbers

[numeric]

A numeric vector to translate.

zero

[character(1)]

What to call values of 0 in numbers (e.g. zero = "zero").

na

[character(1)]

What to call values of NA in numbers (e.g. na = "missing").

nan

[character(1)]

What to call values of NaN in numbers (e.g. nan = "undefined").

inf

[character(1)]

What to call values of Inf in numbers (e.g. inf = "infinity").

negative

[character(1)]

A prefix added to the translation of negative elements of numbers. negative is the string "negative " by default.

decimal

[character(1)]

A word inserted between the whole and fractional part of translated numbers. decimal is the string " and " by default.

and

[TRUE / FALSE]

Whether to insert an " and " before the tens place of translated numbers. and is FALSE by default.

hyphenate

[TRUE / FALSE]

Whether to hyphenate numbers 21 through 99 (e.g. "twenty-one" vs. "twenty one"). hyphenate is TRUE by default.

and_fractional

[TRUE / FALSE]

Whether to insert an " and " before the smallest fractional tens place of translated numbers (e.g. "one hundred one thousandths" vs. "one hundred and one thousandths").

and_fractional is equal to and by default.

hyphenate_fractional

[TRUE / FALSE]

Whether to hyphenate numbers 21 through 99 in the fractional part of translated numbers (e.g. "twenty-one hundredths" or "twenty one hundredths"). This also determines the hyphenation of the fractional units (e.g. "one ten-millionth" vs. "one ten millionth").

hyphenate_fractional is equal to hyphenate by default.

english_fractions

[character]

A named character vector used as a dictionary for the translation of the fractional part of numbers. The names (i.e. keys) are the decimal digits of a fractional number and the values are the corresponding translations.

For example english_fractions = c("5" = "a half") matches the number 0.5 (translated as "a half") and 2.5 (translated as "two and a half").

By default english_fractions is a named character vector with translations for fractions x / y for x = 1, 2, ..., 8 and y = 1, 2, ..., 9. For example, 2 / 3 is translated as "two thirds" and 1 / 2 is translated as "one half".

Provide an empty character to english_fractions to opt out of any such translations. In this case 1 / 2 is translated as "five tenths" instead of "one half".

Value

A non-NA character vector of the same length as numbers.

Examples

numeric_friendly(c(1/3, 0, 0.999, NA, NaN, Inf, -Inf))
#> [1] "one third"                           
#> [2] "zero"                                
#> [3] "nine hundred ninety-nine thousandths"
#> [4] "missing"                             
#> [5] "not a number"                        
#> [6] "infinity"                            
#> [7] "negative infinity"                   

# Specify the translations of "special" numbers
numeric_friendly(c(1, 0, Inf), zero = "none", inf = "all")
#> [1] "one"  "none" "all" 

# Modify the output formatting
frac <- 8765.4321
numeric_friendly(frac)
#> [1] "eight thousand seven hundred sixty-five and four thousand three hundred twenty-one ten-thousandths"
numeric_friendly(frac, decimal = " dot ")
#> [1] "eight thousand seven hundred sixty-five dot four thousand three hundred twenty-one ten-thousandths"
numeric_friendly(frac, hyphenate = TRUE, hyphenate_fractional = FALSE)
#> [1] "eight thousand seven hundred sixty-five and four thousand three hundred twenty one ten thousandths"
numeric_friendly(frac, and = TRUE, and_fractional = TRUE, decimal = " . ")
#> [1] "eight thousand seven hundred and sixty-five . four thousand three hundred and twenty-one ten-thousandths"

# The `friendlynumber.numeric.digits` option specifies the number of
# numeric digits mentioned by `numeric_friendly()`
opts <- options()
options(friendlynumber.numeric.digits = 5)
numeric_friendly(0.0987654321)
#> [1] "nine thousand eight hundred seventy-seven hundred-thousandths"

options(friendlynumber.numeric.digits = 10)
numeric_friendly(0.0987654321)
#> [1] "nine hundred eighty-seven million six hundred fifty-four thousand three hundred twenty-one ten-billionths"
options(opts)

# Set `english_fractions` to specify the translation of certain
# fractions. The names (keys) of `english_fractions` should match
# the decimal part of a fraction (e.g. `"5"` matches `0.5`).
numeric_friendly(
  c(1/2, 6/5, 12),
  english_fractions = c(`5` = "1/2", `2` = "1/5")
)
#> [1] "1/2"         "one and 1/5" "twelve"     

# Input validation
try(numeric_friendly_safe("A"))
#> Error : `numbers` must be numeric, not the string "A".